fix/feat: limpeza de debugs/testes aprimorados correção do fail test & botão de reserva

+++ ViagemImpacta/backend/tests/ViagemImpacta.UnitTests/Helpers/TestHelpers.cs
@@ -41,7 +41,7 @@ public class HotelBuilder
         return this;
     }
 
-    public HotelBuilder WithRating(decimal rating)
+    public HotelBuilder WithRating(float rating)
     {
         _hotel.Rating = rating;
         return this;
